anti slap. hope this one will work now

+++ modules/special_case.js
@@ -10,31 +10,33 @@ module.exports = (global_params,global_settings,message_type) => {
 		yush id: 109385289813798912
 		Arciel id: 151255303852261376
 	*/
-/*	if((message_content.includes("slap") || message_content.includes("punch")) && (message_content.includes("kirimaru")) || message_content.includes("<@"+global_params.env_vars.OWNER_ID+">")){
+	if((message_content.includes("slap") || message_content.includes("punch")) && (message_content.includes("kirimaru")) || message_content.includes("<@"+global_params.env_vars.OWNER_ID+">")){
 		//fetch messages sent by bot after this command
-		global_params.message.channel.fetchMessages({limit:5,after:global_params.message.id}).then(messages => {
-        	const botMessages = messages.filter(msg => msg.author.bot);
-        	global_params.logger.debug(botMessages);
-        	global_params.message.channel.bulkDelete(botMessages).then(function(messages){
-        		var messagesDeleted = botMessages.array().length; 
-        		global_params.message.channel.send("No one is allowed to slap or punch my mastah!!").then(function(){
-	        		global_params.logger.debug("Reply success");	
-	        	}).catch(function(){
-	        		global_params.logger.debug("Failed to send message an error occured");	
-	        	});
-        		global_params.message.delete().then(function(message){
+		setTimeout(function(){
+			global_params.message.channel.fetchMessages({limit:5,after:global_params.message.id}).then(messages => {
+	        	const botMessages = messages.filter(msg => msg.author.bot);
+	        	/*global_params.logger.debug(botMessages);*/
+	        	global_params.message.channel.bulkDelete(botMessages).then(function(messages){
+	        		var messagesDeleted = botMessages.array().length; 
+	        		global_params.message.channel.send("No one is allowed to slap or punch my mastah!!").then(function(){
+		        		global_params.logger.debug("Reply success");	
+		        	}).catch(function(){
+		        		global_params.logger.debug("Failed to send message an error occured");	
+		        	});
+	        		global_params.message.delete().then(function(message){
+		        	}).catch(function(error){
+		        		global_params.logger.debug(error);		
+		        	});   
+		        	console.log("deleted "+messagesDeleted+" messages");
+	        		global_params.logger.info("deleted "+messagesDeleted+" messages");     		
 	        	}).catch(function(error){
-	        		global_params.logger.debug(error);		
-	        	});   
-	        	console.log("deleted "+messagesDeleted+" messages");
-        		global_params.logger.info("deleted "+messagesDeleted+" messages");     		
-        	}).catch(function(error){
-        		global_params.logger.debug(error);	
-        	});
-    	}).catch(function(error){
-    		global_params.logger.debug(error);
-    	});
-	}else */if(global_params.message.author.id === "232221425329504257" && global_settings.f2p === true){
+	        		global_params.logger.debug(error);	
+	        	});
+	    	}).catch(function(error){
+	    		global_params.logger.debug(error);
+	    	});
+		},1000);
+	}else if(global_params.message.author.id === "232221425329504257" && global_settings.f2p === true){
 		/*emi case 232221425329504257 */
 		try{
 			global_params.logger.debug("user matched");
